+++ src/api/auth/kakaoAuth.ts
@@ -1,5 +1,5 @@
-import { instance } from "@/lib/api";
 import { AuthResponse } from "@/types/signup";
+import { instance } from "../api";
 
 export const kakaoAuth = (code: string) => {
   return instance.post<AuthResponse>(
